Move flow functions to workflow
