--- graphrag/index/flows/create_base_text_units.py
@@ -1,119 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""All the steps to transform base text_units."""
-
-import json
-from typing import Any, cast
-
-import pandas as pd
-
-from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
-from graphrag.config.models.chunking_config import ChunkStrategyType
-from graphrag.index.operations.chunk_text.chunk_text import chunk_text
-from graphrag.index.operations.chunk_text.strategies import get_encoding_fn
-from graphrag.index.utils.hashing import gen_sha512_hash
-from graphrag.logger.progress import Progress
-
-
-def create_base_text_units(
-    documents: pd.DataFrame,
-    callbacks: WorkflowCallbacks,
-    group_by_columns: list[str],
-    size: int,
-    overlap: int,
-    encoding_model: str,
-    strategy: ChunkStrategyType,
-    prepend_metadata: bool = False,
-    chunk_size_includes_metadata: bool = False,
-) -> pd.DataFrame:
-    """All the steps to transform base text_units."""
-    sort = documents.sort_values(by=["id"], ascending=[True])
-
-    sort["text_with_ids"] = list(
-        zip(*[sort[col] for col in ["id", "text"]], strict=True)
-    )
-
-    callbacks.progress(Progress(percent=0))
-
-    agg_dict = {"text_with_ids": list}
-    if "metadata" in documents:
-        agg_dict["metadata"] = "first"  # type: ignore
-
-    aggregated = (
-        (
-            sort.groupby(group_by_columns, sort=False)
-            if len(group_by_columns) > 0
-            else sort.groupby(lambda _x: True)
-        )
-        .agg(agg_dict)
-        .reset_index()
-    )
-    aggregated.rename(columns={"text_with_ids": "texts"}, inplace=True)
-
-    def chunker(row: dict[str, Any]) -> Any:
-        line_delimiter = ".\n"
-        metadata_str = ""
-        metadata_tokens = 0
-
-        if prepend_metadata and "metadata" in row:
-            metadata = row["metadata"]
-            if isinstance(metadata, str):
-                metadata = json.loads(metadata)
-            if isinstance(metadata, dict):
-                metadata_str = (
-                    line_delimiter.join(f"{k}: {v}" for k, v in metadata.items())
-                    + line_delimiter
-                )
-
-            if chunk_size_includes_metadata:
-                encode, _ = get_encoding_fn(encoding_model)
-                metadata_tokens = len(encode(metadata_str))
-                if metadata_tokens >= size:
-                    message = "Metadata tokens exceeds the maximum tokens per chunk. Please increase the tokens per chunk."
-                    raise ValueError(message)
-
-        chunked = chunk_text(
-            pd.DataFrame([row]).reset_index(drop=True),
-            column="texts",
-            size=size - metadata_tokens,
-            overlap=overlap,
-            encoding_model=encoding_model,
-            strategy=strategy,
-            callbacks=callbacks,
-        )[0]
-
-        if prepend_metadata:
-            for index, chunk in enumerate(chunked):
-                if isinstance(chunk, str):
-                    chunked[index] = metadata_str + chunk
-                else:
-                    chunked[index] = (
-                        (chunk[0], metadata_str + chunk[1], chunk[2]) if chunk else None
-                    )
-
-        row["chunks"] = chunked
-        return row
-
-    aggregated = aggregated.apply(lambda row: chunker(row), axis=1)
-
-    aggregated = cast("pd.DataFrame", aggregated[[*group_by_columns, "chunks"]])
-    aggregated = aggregated.explode("chunks")
-    aggregated.rename(
-        columns={
-            "chunks": "chunk",
-        },
-        inplace=True,
-    )
-    aggregated["id"] = aggregated.apply(
-        lambda row: gen_sha512_hash(row, ["chunk"]), axis=1
-    )
-    aggregated[["document_ids", "chunk", "n_tokens"]] = pd.DataFrame(
-        aggregated["chunk"].tolist(), index=aggregated.index
-    )
-    # rename for downstream consumption
-    aggregated.rename(columns={"chunk": "text"}, inplace=True)
-
-    return cast(
-        "pd.DataFrame", aggregated[aggregated["text"].notna()].reset_index(drop=True)
-    )
