--- graphrag/index/flows/create_community_reports_text.py
@@ -1,67 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""All the steps to transform community reports."""
-
-import logging
-
-import pandas as pd
-
-from graphrag.cache.pipeline_cache import PipelineCache
-from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
-from graphrag.config import defaults
-from graphrag.config.enums import AsyncType
-from graphrag.index.operations.finalize_community_reports import (
-    finalize_community_reports,
-)
-from graphrag.index.operations.summarize_communities.explode_communities import (
-    explode_communities,
-)
-from graphrag.index.operations.summarize_communities.summarize_communities import (
-    summarize_communities,
-)
-from graphrag.index.operations.summarize_communities.text_unit_context.context_builder import (
-    build_level_context,
-    build_local_context,
-)
-
-log = logging.getLogger(__name__)
-
-
-async def create_community_reports_text(
-    entities: pd.DataFrame,
-    communities: pd.DataFrame,
-    text_units: pd.DataFrame,
-    callbacks: WorkflowCallbacks,
-    cache: PipelineCache,
-    summarization_strategy: dict,
-    async_mode: AsyncType = AsyncType.AsyncIO,
-    num_threads: int = 4,
-) -> pd.DataFrame:
-    """All the steps to transform community reports."""
-    nodes = explode_communities(communities, entities)
-
-    summarization_strategy["extraction_prompt"] = summarization_strategy["text_prompt"]
-
-    max_input_length = summarization_strategy.get(
-        "max_input_length", defaults.COMMUNITY_REPORT_MAX_INPUT_LENGTH
-    )
-
-    local_contexts = build_local_context(
-        communities, text_units, nodes, max_input_length
-    )
-
-    community_reports = await summarize_communities(
-        nodes,
-        communities,
-        local_contexts,
-        build_level_context,
-        callbacks,
-        cache,
-        summarization_strategy,
-        max_input_length=max_input_length,
-        async_mode=async_mode,
-        num_threads=num_threads,
-    )
-
-    return finalize_community_reports(community_reports, communities)
