--- graphrag/index/flows/finalize_graph.py
@@ -1,26 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""All the steps to create the base entity graph."""
-
-import pandas as pd
-
-from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
-from graphrag.config.models.embed_graph_config import EmbedGraphConfig
-from graphrag.index.operations.finalize_entities import finalize_entities
-from graphrag.index.operations.finalize_relationships import finalize_relationships
-
-
-def finalize_graph(
-    entities: pd.DataFrame,
-    relationships: pd.DataFrame,
-    callbacks: WorkflowCallbacks,
-    embed_config: EmbedGraphConfig | None = None,
-    layout_enabled: bool = False,
-) -> tuple[pd.DataFrame, pd.DataFrame]:
-    """All the steps to finalize the entity and relationship formats."""
-    final_entities = finalize_entities(
-        entities, relationships, callbacks, embed_config, layout_enabled
-    )
-    final_relationships = finalize_relationships(relationships)
-    return (final_entities, final_relationships)
