+++ graphrag/index/workflows/extract_covariates.py
@@ -3,11 +3,18 @@
 
 """A module containing run_workflow method definition."""
 
+from typing import Any
+from uuid import uuid4
+
+import pandas as pd
+
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
+from graphrag.config.enums import AsyncType
 from graphrag.config.models.graph_rag_config import GraphRagConfig
 from graphrag.index.context import PipelineRunContext
-from graphrag.index.flows.extract_covariates import (
-    extract_covariates,
+from graphrag.index.operations.extract_covariates.extract_covariates import (
+    extract_covariates as extractor,
 )
 from graphrag.index.typing import WorkflowFunctionOutput
 from graphrag.utils.storage import load_table_from_storage, write_table_to_storage
@@ -47,3 +54,51 @@ async def run_workflow(
     await write_table_to_storage(output, "covariates", context.storage)
 
     return WorkflowFunctionOutput(result=output, config=None)
+
+
+async def extract_covariates(
+    text_units: pd.DataFrame,
+    callbacks: WorkflowCallbacks,
+    cache: PipelineCache,
+    covariate_type: str,
+    extraction_strategy: dict[str, Any] | None,
+    async_mode: AsyncType = AsyncType.AsyncIO,
+    entity_types: list[str] | None = None,
+    num_threads: int = 4,
+) -> pd.DataFrame:
+    """All the steps to extract and format covariates."""
+    # reassign the id because it will be overwritten in the output by a covariate one
+    # this also results in text_unit_id being copied to the output covariate table
+    text_units["text_unit_id"] = text_units["id"]
+    covariates = await extractor(
+        input=text_units,
+        callbacks=callbacks,
+        cache=cache,
+        column="text",
+        covariate_type=covariate_type,
+        strategy=extraction_strategy,
+        async_mode=async_mode,
+        entity_types=entity_types,
+        num_threads=num_threads,
+    )
+    text_units.drop(columns=["text_unit_id"], inplace=True)  # don't pollute the global
+    covariates["id"] = covariates["covariate_type"].apply(lambda _x: str(uuid4()))
+    covariates["human_readable_id"] = covariates.index + 1
+
+    return covariates.loc[
+        :,
+        [
+            "id",
+            "human_readable_id",
+            "covariate_type",
+            "type",
+            "description",
+            "subject_id",
+            "object_id",
+            "status",
+            "start_date",
+            "end_date",
+            "source_text",
+            "text_unit_id",
+        ],
+    ]
