+++ graphrag/index/workflows/finalize_graph.py
@@ -3,13 +3,15 @@
 
 """A module containing run_workflow method definition."""
 
+import pandas as pd
+
 from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
+from graphrag.config.models.embed_graph_config import EmbedGraphConfig
 from graphrag.config.models.graph_rag_config import GraphRagConfig
 from graphrag.index.context import PipelineRunContext
-from graphrag.index.flows.finalize_graph import (
-    finalize_graph,
-)
 from graphrag.index.operations.create_graph import create_graph
+from graphrag.index.operations.finalize_entities import finalize_entities
+from graphrag.index.operations.finalize_relationships import finalize_relationships
 from graphrag.index.operations.snapshot_graphml import snapshot_graphml
 from graphrag.index.typing import WorkflowFunctionOutput
 from graphrag.utils.storage import load_table_from_storage, write_table_to_storage
@@ -53,3 +55,18 @@ async def run_workflow(
         },
         config=None,
     )
+
+
+def finalize_graph(
+    entities: pd.DataFrame,
+    relationships: pd.DataFrame,
+    callbacks: WorkflowCallbacks,
+    embed_config: EmbedGraphConfig | None = None,
+    layout_enabled: bool = False,
+) -> tuple[pd.DataFrame, pd.DataFrame]:
+    """All the steps to finalize the entity and relationship formats."""
+    final_entities = finalize_entities(
+        entities, relationships, callbacks, embed_config, layout_enabled
+    )
+    final_relationships = finalize_relationships(relationships)
+    return (final_entities, final_relationships)
